requirements.txt
benepar==0.2.0
click==7.1.2
conceptnet_lite==0.1.27
dropbox==11.30.0
fitter==1.4.0
mlflow==1.22.0
negspacy==1.0.2
networkx==2.7.1
nltk==3.6.5
numpy==1.22.0
pandarallel==1.6.1
pandas==1.3.5
pingouin==0.5.1
plotly==5.5.0
py2neo==2021.2.3
PyMultiDictionary==1.2.0
python_igraph==0.9.10
pytorch_lightning==1.5.5
PyYAML==6.0
ray==1.9.2
requests==2.26.0
sacremoses==0.0.47
scikit_learn==1.1.1
scipy==1.7.3
sentence_transformers==2.1.0
setuptools==57.0.0
spacy==3.2.1
statsmodels==0.13.2
streamlit==1.7.0
textblob==0.17.1
textstat==0.7.2
tokenizers==0.10.3
torch==1.10.0
torchmetrics==0.6.1
tqdm==4.62.3
transformers==4.13.0
